requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via httpx
apscheduler==3.11.1
    # via python-telegram-bot
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via python-telegram-bot
idna==3.11
    # via
    #   anyio
    #   httpx
pydantic==2.12.5
    # via -r requirements.in
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.0.0
    # via -r requirements.in
python-telegram-bot[job-queue]==22.5
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   anyio
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via tzlocal
tzlocal==5.3.1
    # via apscheduler
